requirements.txt
# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model evaluation and metrics
imbalanced-learn>=0.11.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Database
pymongo>=4.6.0
python-dotenv>=1.0.0
